refactor : AuthUiState 클래스 분리
