+++ app/src/main/java/com/example/assignmenttest/AuthViewModel.kt
@@ -89,13 +89,3 @@ class AuthViewModel @Inject constructor(
     }
 
 }
-
-
-sealed class AuthUiState {
-    object Idle : AuthUiState()
-    object Loading : AuthUiState()
-    object Success : AuthUiState()
-    object Logout : AuthUiState()
-    object AccountDelete : AuthUiState()
-    data class Error(val message: String) : AuthUiState()
-}
